+++ Examples/AdvancedExample02.ps1
@@ -188,7 +188,7 @@ $advancedexample02 = & {
         This is not possible with native Graphviz SubGraphs.\n'
         Add-DiaNodeText -Name 'Info-Box2' -DraftMode:$DraftMode -Text $Message -FontColor '#0000FF' -FontSize 20 -FontBold -TextAlign 'Left' -TableBorder 0 -NodeObject
 
-        Edge -From 'USA-WebServers' -To 'Info-Box2' -Attributes @{minlen = 2; color = 'black'; style = 'dashed'; penwidth = 1.5 }
+        Add-DiaNodeEdge -From 'USA-WebServers' -To 'Info-Box2' -EdgeLength 2 -EdgeColor 'black' -EdgeStyle 'dashed' -GraphvizAttributes @{penwidth = 1.5}
     }
 }
 <#