+++ CHANGELOG.md
@@ -5,6 +5,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.39] - 2026-03-02
+
+### :toolbox: Added
+
+- Add a cmdlet to the DiaConvertImageToPDF .Net Assembly (New-ConvertPngToPdf)
+- Add a cmdlet to the Diagrammer .Net Assembly:
+  - New-ConvertPngToPdf
+  - New-WatermarkToImage
+  - Get-ImageHeightFromBase64
+  - Get-ImageHeightFromFile
+  - Get-ImagePropertiesFromFile
+  - Get-ImageWidthFromBase64
+  - Get-ImageWidthFromFile
+  - Get-ResizeImageFromFile
+  - Get-RotateImageFromFile
+- Add function to add edges between nodes
+  - (Add-DiaNodeEdge)
+
+### :arrows_clockwise: Changed
+
+- Upgrade assembly libraries
+- Update module version to v0.2.39
+
 ## [0.2.38] - 2026-02-21
 
 ### :arrows_clockwise: Changed